pyproject.toml
[tool.poetry]
name = "hemm"
version = "0.1.0"
description = ""
authors = ["soumik12345 <19soumik.rakshit96@gmail.com>"]
readme = "README.md"
packages = [{ include = "hemm" }]

[tool.poetry.dependencies]
python = "^3.10"
diffusers = "^0.27.2"
accelerate = "^0.30.1"
wandb = "^0.17.0"
poetry = "^1.8.3"
weave = "^0.50.9"
huggingface-hub = "^0.23.0"
datasets = "^2.19.1"
fire = "^0.6.0"
mkdocs = "^1.4.3"
mkdocstrings = { extras = ["python"], version = "^0.22.0" }
mkdocs-material = "^9.1.15"
mkdocs-minify-plugin = "^0.6.4"
mkdocs-glightbox = "^0.3.4"
mkdocs-jupyter = "^0.24.1"
jupyter = "^1.0.0"
black = "^24.4.2"
torcheval = "^0.0.7"
ruff = "^0.4.5"
openai = "^1.33.0"
transformers = "^4.41.2"
jsonlines = "^4.0.0"
spacy = "^3.7.5"
instructor = "^1.3.4"
torchmetrics = { extras = ["multimodal"], version = "^1.4.1" }

[tool.poetry.extras]
core = [
    "diffusers",
    "transformers",
    "accelerate",
    "wandb",
    "weave",
    "poetry",
    "huggingface-hub",
    "datasets",
    "fire",
    "jsonlines",
    "spacy",
    "instructor",
    "torchmetrics",
]
docs = [
    "mkdocs",
    "mkdocstrings",
    "mkdocs-material",
    "mkdocs-minify-plugin",
    "mkdocs-glightbox",
    "mkdocs-jupyter",
    "jupyter",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"